requirements.txt
pandas
matplotlib
Pillow
opencv-python
numpy